requirements.txt
# Flask and Server
Flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.9

# Authentication
PyJWT==2.8.0

# HTTP Requests
requests==2.31.0

# Environment
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Code Quality
black==23.11.0
flake8==6.1.0
pylint==3.0.3
